Changes made up to 11/27/24 11:08
Mostly minor changes. Fixed the hydration error, which would cause the two-way and event bindings not to work on other components outside of the main file. (body tag can only be used in the index.HTML).

Created a mock-heroes ts file in order to store data there. The array is of type heroes (from heroes.ts, the interface) and its then imported to heroes component so we can loop thru and display it in using the template file. (it is defined in the ts file --> heroes = HEROES; )

Made it so that the entire page is one background color. Referring to earlier, the body tag outside of index.html was causing errors with hydration, which is why the user-side to server-side actions would never work, but the ones from the server side to the client side always worked. (there were never any terminal errors or warnings, but there were errors in the terminal).
I did this by going into the global styles page (styles.css) and setting html { background-color: ...; }
This set the background color to the one I set for the entire web app. I may have to change this in the future if needed, but using the body tag outside of index.html will cause hydration issues.