Changes [3/27/24, 3:40]
Minor changes are following along.

Created event binding of function onSelect() which takes in input of hero. Assuming "hero" is the instance of the interface which we imported called Hero. We put id:1, name: 'Windstorm' , summary: 'This is a template summary of the hero'

The function onSelect() takes in an arguement hero of type Hero. Where Hero is the general interface we imported and hero was the specific one with id:1, name: windstorm, etc. Then within the function we call for the specific instance of selected hero (which outside the function is defined as selectedHero?:Hero) I dont know what that means but I am assuming that its asking if its type is Hero [?]

Remember that Hero is the interface which we imported with those 3 different category. I am guessing that this line is saying that instance of the selectedHero. is set equal to the hero which is selected [?]

Then I have an if statement saying if "selectedHero" is selected then it shows the attributes below such as names, ids.

I have a two-way binding where the user can edit it, and it makes changes to the buttons as well. I use selectedHero.name which targets the name attribute of selectedHero